requirements.txt
functions-framework==3.*
google-cloud-bigquery
pytz~=2025.2
protobuf~=6.31.1
Flask~=3.1.1